login logout

+++ Raw/Python/study_budy/base/views.py
@@ -1,5 +1,11 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -9,6 +15,40 @@ from .forms import RoomForm
 #     {"id": 2, "name": "Design with me"},
 #     {"id": 3, "name": "Frontend developers"},
 # ] # <- rooms = Room.objects.all()
+
+
+def login_page(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
+    if request.method == "POST":
+        username = request.POST.get("username")
+        password = request.POST.get("password")
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "Username or password does not exist")
+    context = {"page": page}
+    return render(request, "base/login_register.html", context)
+
+
+def logout_user(request):
+    logout(request)
+    return redirect("home")
+
+
+def register_page(request):
+    page = "register"
+    return render(request, "base/login_register.html")
 
 
 def home(request):
@@ -30,6 +70,7 @@ def room(request, pk):
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="login")
 def create_room(request):
     form = RoomForm
     if request.method == "POST":
@@ -41,9 +82,13 @@ def create_room(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to do this!")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -54,8 +99,13 @@ def update_room(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to do this!")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
